requirements.txt
pymongo==4.12.0
langchain==0.3.18
langchain-community==0.3.16
pypdf==5.4.0
python-dotenv==1.0.1
openai==1.74.0
einops==0.8.1
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.11.3
sentence-transformers==4.1.0
python-multipart==0.0.20
pinecone-client==5.4.2
pdfplumber==0.11.5
